Corrigido o tipo de retorno em Usuario.getChaveImagemPerfil() para long

+++ app/src/main/java/com/infact/nightour/model/Usuario.java
@@ -114,7 +114,7 @@ public class Usuario {
         this.aniversario = aniversario;
     }
 
-    public int getChaveImagemPerfil() {
+    public long getChaveImagemPerfil() {
         return getImagemPerfil().getId();
     }
